The auth0 library declares types that are out of date, workaround this.
Siddharth Kshetrapal was contacted on what to do about the auth0
typescript file causing issues because it declares several invalid
types.
He mentioned that they no longer use typescript with their components,
instead preferring to generate typings for vscode users.
Assuming it's fine to just delete the file and move on until auth0 fixes
this, I'm running the deletion of that file after any 'npm install' with
the postinstall builtin script.

To this end, it was achieved with adding a lib that can delete files
'del-cli' and a lib that can control which scripts are run and how,
npm-run-all.
run-all is used to run-s, meaning, run sequentially, deletion of the
problematic file .